Improve amd64 JIT helpers and regression harness

+++ tests/manual/atomic_cmpxchg32.c
@@ -15,7 +15,7 @@ struct worker_result {
 
 static inline int lock_cmpxchg_u32(volatile uint32_t *ptr, uint32_t expected,
                                    uint32_t desired, uint32_t *actual_out) {
-    uint32_t eflags;
+    unsigned long eflags;
     asm volatile(
         "lock cmpxchgl %4, %1\n\t"
         "pushf\n\t"
@@ -31,7 +31,7 @@ static void test_lock_cmpxchgl_single(uint32_t eax_init) {
     volatile uint32_t mem = 0xfbca7654u;
     uint32_t eax = eax_init;
     uint32_t src = 0x12345678u;
-    uint32_t eflags;
+    unsigned long eflags;
     uint32_t expected_result = eax_init - mem;
     uint32_t expected_flags = sub_flags32(eax_init, mem, expected_result);
     uint32_t expected_mem = eax_init == mem ? src : mem;
@@ -91,7 +91,7 @@ static void test_lock_cmpxchgl_vectors(void) {
         uint32_t eax = next_u32(&seed);
         uint32_t initial_eax = eax;
         uint32_t src = next_u32(&seed);
-        uint32_t eflags;
+        unsigned long eflags;
         uint32_t result = initial_eax - initial_mem;
         uint32_t expected_flags = sub_flags32(initial_eax, initial_mem, result);
         uint32_t expected_mem = initial_eax == initial_mem ? src : initial_mem;
